@@ -2,7 +2,7 @@ name: univibe
 description: "An University Event Tracking System."
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -32,39 +32,39 @@ dependencies:
     sdk: flutter
 
   # Supabase
-  supabase_flutter: ^1.10.0
-  
+  supabase_flutter: ^2.12.0
+
   # State Management
   provider: ^6.1.0
-  
+
   # Navigation
-  go_router: ^12.0.0
-  
+  go_router: ^17.0.1
+
   # UI & Icons
   cupertino_icons: ^1.0.8
-  google_fonts: ^6.0.0
+  google_fonts: ^7.1.0
   flutter_svg: ^2.0.7
   shimmer: ^3.0.0
-  
+
   # Date & Time
-  intl: ^0.19.0
-  
+  intl: ^0.20.2
+
   # Image Handling
   image_picker: ^1.0.0
   cached_network_image: ^3.3.0
-  
+
   # Calendar & Notifications
   table_calendar: ^3.0.0
-  flutter_local_notifications: ^15.0.0
-  timezone: ^0.9.0
-  
+  flutter_local_notifications: ^19.5.0
+  timezone: ^0.10.1
+
   # Location
-  geolocator: ^10.0.0
+  geolocator: ^14.0.2
   url_launcher: ^6.2.0
-  
+
   # Storage
   shared_preferences: ^2.2.0
-  
+
   # Utils
   uuid: ^4.0.0
   http: ^1.1.0
@@ -79,17 +79,24 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^6.0.0
+  flutter_launcher_icons: ^0.13.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
 # The following section is specific to Flutter packages.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
+
+# Flutter Launcher Icons configuration
+flutter_launcher_icons:
+  android: "ic_launcher"
+  image_path: "assets/app_icon.png"
+  min_sdk_android: 21 # Required for adaptive icons
+
 
   # To add assets to your application, add an assets section, like this:
   # assets:
