@@ -83,6 +83,11 @@ class _LoginScreenState extends State<LoginScreen> {
                       const SizedBox(height: 8),
                       TextField(
                         controller: _emailController,
+                        style: TextStyle(
+                          color: Theme.of(context).brightness == Brightness.dark
+                              ? Colors.white
+                              : Colors.black,
+                        ),
                         decoration: InputDecoration(
                           prefixIcon: Icon(
                             Icons.email,
@@ -90,7 +95,10 @@ class _LoginScreenState extends State<LoginScreen> {
                           ),
                           hintText: 'Enter your email',
                           filled: true,
-                          fillColor: Colors.grey[100],
+                          fillColor:
+                              Theme.of(context).brightness == Brightness.dark
+                              ? Colors.grey[800]
+                              : Colors.grey[100],
                           border: OutlineInputBorder(
                             borderRadius: BorderRadius.circular(12),
                             borderSide: BorderSide.none,
@@ -111,6 +119,11 @@ class _LoginScreenState extends State<LoginScreen> {
                       const SizedBox(height: 8),
                       TextField(
                         controller: _passwordController,
+                        style: TextStyle(
+                          color: Theme.of(context).brightness == Brightness.dark
+                              ? Colors.white
+                              : Colors.black,
+                        ),
                         decoration: InputDecoration(
                           prefixIcon: Icon(
                             Icons.lock,
@@ -131,7 +144,10 @@ class _LoginScreenState extends State<LoginScreen> {
                           ),
                           hintText: 'Enter your password',
                           filled: true,
-                          fillColor: Colors.grey[100],
+                          fillColor:
+                              Theme.of(context).brightness == Brightness.dark
+                              ? Colors.grey[800]
+                              : Colors.grey[100],
                           border: OutlineInputBorder(
                             borderRadius: BorderRadius.circular(12),
                             borderSide: BorderSide.none,
